+++ changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Next
-- Added jagged terrain generation
+- New save file format (v2.0) using run-length encoding
+  - Saves roughly 500x file size and makes saving and loading much faster
 - Fixed input not working after losing focus
 
 ## 0.1.3_1